compiler: [major] refactor C-gen code
