+++ cogent/src/Cogent/Core.hs
@@ -198,7 +198,7 @@ traverseE f (Unit)               = pure $ Unit
 traverseE f (ILit i pt)          = pure $ ILit i pt
 traverseE f (SLit s)             = pure $ SLit s
 traverseE f (ALit es)            = ALit <$> traverse f es
-traverseE f (Let a e1 e2)        = Let a  <$> f e1 <*> f e2
+traverseE f (Let a e1 e2)        = Let a <$> f e1 <*> f e2
 traverseE f (LetBang vs a e1 e2) = LetBang vs a <$> f e1 <*> f e2
 traverseE f (Tuple e1 e2)        = Tuple <$> f e1 <*> f e2
 traverseE f (Struct fs)          = Struct <$> traverse (traverse f) fs
