+++ cogent/src/Cogent/Haskell/HscSyntax.hs
@@ -1,13 +1,14 @@
 --
--- Copyright 2017, NICTA
+-- Copyright 2018, Data61
+-- Commonwealth Scientific and Industrial Research Organisation (CSIRO)
+-- ABN 41 687 119 230.
 --
 -- This software may be distributed and modified according to the terms of
 -- the GNU General Public License version 2. Note that NO WARRANTY is provided.
 -- See "LICENSE_GPLv2.txt" for details.
 --
--- @TAG(NICTA_GPL)
+-- @TAG(DATA61_GPL)
 --
-
 
 module Cogent.Haskell.HscSyntax where
 
@@ -26,7 +27,8 @@ data ModulePragma = LanguagePragma String
 
 data Declaration = HsDecl HsDecl | HscDecl HscDecl | EmptyDecl
 
-data HsDecl = ImportDecl ModuleName Qualified (Maybe ModuleName) [VarName] [VarName]  -- ImportDecl name is-qualified? short-name include-list exclude-list
+data HsDecl = ImportDecl ModuleName Qualified (Maybe ModuleName) [VarName] [VarName]
+              -- ^ @ImportDecl name is-qualified? short-name include-list exclude-list@
             | DataDecl TypeName [TyVarName] [DataCon]
             | TypeDecl TypeName [TyVarName] Type
             | InstDecl ClassName [Context] Type [Binding]
